refactor code and make first passing tests
