+++ storage/core.ts
@@ -1,0 +1,5 @@
+import { join } from 'path'
+
+export const getLANStoragePath = (baseDir: string): string => {
+    return join(baseDir, 'sltt-app', 'lan')
+}
