+++ storage/test-data/listTests/local-and-remote/testProject/local/local-doc__2024-07-25_14-50-23-046__210629_180535-240725_145023__c62114c2__c62114c2.sltt-doc
@@ -1,0 +1,1 @@
+{"_id":"210629_180535/240725_145023","assignee":"","contentType":"Translation","creationDate":"2024/07/25 14:50:23.046Z","creator":"epyle@biblesocieties.org","difficulty":1,"modBy":"epyle@biblesocieties.org","modDate":"2024/07/25 14:50:23.046Z","name":"test offline passage","rank":"00000001000.0000","references":"[]"}
