+++ storage/test-data/listTests/local-and-remote/testProject/remote/000066977__2024-07-25_16-26-36-672__210202_183235-240607_145904-240618_160543-240725_162634__e85c7697__e85c7697.sltt-doc
@@ -1,0 +1,1 @@
+{"_id":"210202_183235/240607_145904/240618_160543/240725_162634","canResolve":true,"creationDate":"2024/07/25 16:26:34.941Z","creator":"nmiles@biblesocieties.org","description":"","endPosition":2.976,"modBy":"nmiles@biblesocieties.org","modDate":"2024/07/25 16:26:36.672Z","position":1.3862099109911736,"rank":"00000000001.3862","startPosition":0,"type":"0"}
